lotline-intelligence/scripts/requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pandas>=2.2.0
geopandas>=0.14.0
shapely>=2.0.0
sqlalchemy>=2.0.0
schedule>=1.2.1
